+++ src/main/java/woowacourse/member/dto/request/DuplicateEmailRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.member.dto;
+package woowacourse.member.dto.request;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
