+++ src/main/java/woowacourse/member/dto/response/MemberInfoResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.member.dto;
+package woowacourse.member.dto.response;
 
 import woowacourse.member.domain.Member;
