+++ src/test/java/woowacourse/auth/acceptance/AuthAcceptanceTest.java
@@ -5,9 +5,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import woowacourse.acceptance.AcceptanceTest;
 import woowacourse.acceptance.RestAssuredConvenienceMethod;
-import woowacourse.auth.dto.LoginRequest;
+import woowacourse.member.dto.request.LoginRequest;
 import woowacourse.auth.dto.TokenResponse;
-import woowacourse.member.dto.SignUpRequest;
+import woowacourse.member.dto.request.SignUpRequest;
 
 import static org.hamcrest.Matchers.equalTo;
